requirements.txt
flake8==7.3.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
numpy==2.2.6
pillow==11.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
rich==14.1.0
tqdm==4.67.1
